Added max_delta
If the values of the sensors change too fast, there might be an issue with the sensor. Mine are crap like that and could consistently just show 10 C higher temperatures all of a sudden. Powering the sensor down knocks some sense back into it and it starts showing the actual temperatures again.